refactor: service 로직에서 http 호출하도록 변경
